UI improvements and bug fixes

+++ backend/src/main/java/com/viniciusmcabral/sound_rate/dtos/response/AlbumRatingDTO.java
@@ -1,5 +1,4 @@
 package com.viniciusmcabral.sound_rate.dtos.response;
 
-public record AlbumRatingDTO(Long id, Integer rating, UserDTO author) {
-
+public record AlbumRatingDTO(Long id, Double rating, UserDTO author) {
 }
