+++ backend/src/main/java/com/viniciusmcabral/sound_rate/dtos/response/TrackRatingDTO.java
@@ -1,5 +1,4 @@
 package com.viniciusmcabral.sound_rate.dtos.response;
 
-public record TrackRatingDTO(Long id, Integer rating, String trackId, UserDTO author) {
-
+public record TrackRatingDTO(Long id, Double rating, String trackId, UserDTO author) {
 }
